requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
asyncpg==0.29.0

# AI/ML Libraries
torch==2.1.1
transformers==4.36.0
tokenizers==0.15.0
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4

# Vector Database
chromadb==0.4.18
faiss-cpu==1.7.4

# API Integration
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0
openai==1.3.7
anthropic==0.7.7

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
loguru==0.7.2

# Task Queue & Background Jobs
celery==5.3.4
redis==5.0.1

# Web Framework & UI
jinja2==3.1.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1
pyyaml==6.0.1
toml==0.10.2

# Plugin System
importlib-metadata==6.8.0
pluggy==1.3.0

# Multilingual Support
polyglot==16.7.4
langdetect==1.0.9

# Data Processing
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Configuration Management
dynaconf==3.2.4

# Async Support
asyncio-mqtt==0.16.1
websockets==12.0

# File Processing
python-magic==0.4.27
pillow==10.1.0

# Geospatial
geopy==2.4.1

# Time & Date
python-dateutil==2.8.2
pytz==2023.3

# Compression
gzip==3.12
lz4==4.3.2

# Network
aiofiles==23.2.1
aiohttp-cors==0.7.0

# Performance
uvloop==0.19.0
orjson==3.9.10 